+++ lib/poetry/core/preview/sidecarable.rb
@@ -61,7 +61,6 @@ module Poetry
         #
         # @param base [Class] the class being extended with this module
         # @return [void]
-        # @api private
         def self.extended(base)
           base.singleton_class.prepend(ClassMethods)
         end
@@ -129,6 +128,8 @@ module Poetry
             name.sub(/(::Preview|Preview)$/, "").underscore
           end
         end
+
+        private_class_method :extended
       end
     end
   end